1st task and readme update

+++ 0x07-pointers_arrays_strings/0-memset.c
@@ -13,7 +13,7 @@ char *_memset(char *s, char b, unsigned int n)
 {
 	int i = 0;
 
-	for (; i < n; i++)
+	for (; i < (int)n; i++)
 	{
 		*(s + i) = b;
 	}
